Fix watch layout bug
The watch command is running for a prolonged period but the clock used
to calculate the snapshot layout is from when dsnapshot started. This
resulted in the new snapshots always being outside of the layout (before
bucket zero).

This in turn meant that they all ended up in the layouts discarded list
and where removed.

This fixes it so an update of the layout is always from the current
clock.